OrderPlacement_Automation

+++ src/com/hk/util/Xls_Reader.java
@@ -157,7 +157,7 @@ public class Xls_Reader {
 		  return cell.getStringCellValue();
 	  else if(cell.getCellType()==Cell.CELL_TYPE_NUMERIC || cell.getCellType()==Cell.CELL_TYPE_FORMULA ){
 		  
-		  String cellText  = String.valueOf(cell.getNumericCellValue());
+		  String cellText  = String.valueOf((long)cell.getNumericCellValue());
 		  if (HSSFDateUtil.isCellDateFormatted(cell)) {
 	           // format in form of M/D/YY
 			  double d = cell.getNumericCellValue();
